requirements.txt
langchain
openai
stratz
gradio
gunicorn
python-dotenv